Some styles change.
I had ghostty and nvim to be configured in a way to have both working
for dark and light themes, however I can't really do that because lsd
and vivid and lazygit get annoying because they don't support double
colors.

Maybe someday I will have patient to check that, but it's not today.